Don't generate contexts of contexts
For the hack in Neovim we need to activate the popup window to clear
highlights. This triggered an autocmd which would try to generate the
context of this popup window.

This lead to other issues like overwriting the previous window we want
to switch back to and also confuses the code which checks for number
column width etc.

fixup! Add workaround to not get stuck in popup sometimes